+++ lib/webpacker/dev_server.rb
@@ -10,8 +10,12 @@ class Webpacker::DevServer
   end
 
   def running?
-    Socket.tcp(host, port, connect_timeout: connect_timeout).close
-    true
+    if config.dev_server.present?
+      Socket.tcp(host, port, connect_timeout: connect_timeout).close
+      true
+    else
+      false
+    end
   rescue
     false
   end