+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 **Please note that Webpacker 3.1.0 and 3.1.1 has some serious bugs so please consider using either 3.0.2 or 3.2.0**
 
+## [Unreleased]
+
+- Disable dev server running? check if no dev server config is present in that environment [#1179](https://github.com/rails/webpacker/pull/1179)
+
 ## [3.2.0] - 2017-12-16
 
 ### To upgrade: